Remove use of context.
