+++ js/src/components/google-combo-account-card/connected-google-combo-account-card.js
@@ -1,9 +1,13 @@
 /**
+ * External dependencies
+ */
+import { useEffect, useRef } from '@wordpress/element';
+
+/**
  * Internal dependencies
  */
 import AccountCard, { APPEARANCE } from '../account-card';
 import AccountCardDescription, { Indicator } from './account-card-description';
-import { AccountCreationContext } from './account-creation-context';
 import AppSpinner from '../app-spinner';
 import useAutoCreateAdsMCAccounts from '.~/hooks/useAutoCreateAdsMCAccounts';
 import useGoogleAdsAccount from '.~/hooks/useGoogleAdsAccount';
@@ -15,14 +19,48 @@ import './connected-google-combo-account-card.scss';
  * It will also kickoff Ads and Merchant Center account creation if the user does not have accounts.
  */
 const ConnectedGoogleComboAccountCard = () => {
-	const { hasFinishedResolution: hasFinishedResolutionForCurrentAdsAccount } =
-		useGoogleAdsAccount();
+	const {
+		googleAdsAccount,
+		hasFinishedResolution: hasFinishedResolutionForCurrentAdsAccount,
+	} = useGoogleAdsAccount();
 
-	const { hasFinishedResolution: hasFinishedResolutionForCurrentMCAccount } =
-		useGoogleMCAccount();
+	const {
+		googleMCAccount,
+		hasFinishedResolution: hasFinishedResolutionForCurrentMCAccount,
+	} = useGoogleMCAccount();
 
 	const { accountsCreated, hasDetermined, creatingWhich } =
 		useAutoCreateAdsMCAccounts();
+
+	const wasCreatingAccounts = useRef( null );
+	const accountDetailsResolved =
+		hasFinishedResolutionForCurrentAdsAccount &&
+		hasFinishedResolutionForCurrentMCAccount;
+
+	useEffect( () => {
+		if ( creatingWhich ) {
+			wasCreatingAccounts.current = creatingWhich;
+		}
+
+		if (
+			wasCreatingAccounts.current &&
+			accountsCreated &&
+			accountDetailsResolved
+		) {
+			const accountsReady =
+				!! googleAdsAccount?.id && !! googleMCAccount?.id;
+
+			if ( accountsReady ) {
+				wasCreatingAccounts.current = null;
+			}
+		}
+	}, [
+		accountDetailsResolved,
+		accountsCreated,
+		creatingWhich,
+		googleAdsAccount,
+		googleMCAccount,
+	] );
 
 	if (
 		! accountsCreated &&
@@ -33,21 +71,20 @@ const ConnectedGoogleComboAccountCard = () => {
 		return <AccountCard description={ <AppSpinner /> } />;
 	}
 
-	const accountCreationData = {
-		accountsCreated,
-		creatingWhich,
-	};
-
 	return (
-		<AccountCreationContext.Provider value={ accountCreationData }>
-			<AccountCard
-				appearance={ APPEARANCE.GOOGLE }
-				alignIcon="top"
-				className="gla-google-combo-account-card--connected"
-				helper={ <AccountCardDescription /> }
-				indicator={ <Indicator /> }
-			/>
-		</AccountCreationContext.Provider>
+		<AccountCard
+			appearance={ APPEARANCE.GOOGLE }
+			alignIcon="top"
+			className="gla-google-combo-account-card--connected"
+			helper={
+				<AccountCardDescription
+					creatingAccounts={ wasCreatingAccounts.current }
+				/>
+			}
+			indicator={
+				<Indicator creatingAccounts={ wasCreatingAccounts.current } />
+			}
+		/>
 	);
 };
